organize files in applications/contrastive_phenotyping
