+++ applications/contrastive_phenotyping/examples_cli/fit_slurm.sh
