+++ applications/contrastive_phenotyping/examples_cli/plot_embeddings.py
