+++ applications/contrastive_phenotyping/examples_cli/predict_slurm.sh
