+++ applications/contrastive_phenotyping/figures/figure_4a_1.py
