+++ applications/contrastive_phenotyping/figures/figure_4e_2_june.py
